Move option processor to the common lib
