+++ common/Cargo.toml
@@ -7,3 +7,4 @@ authors = ["amanjpro <http://www.amanj.me>"]
 serde = "1.0"
 serde_derive = "1.0"
 rust-ini = "0.10"
+docopt = "0.8"
