+++ main/Cargo.toml
@@ -12,7 +12,6 @@ name = "smtpc"
 path = "src/smtpc.rs"
 
 [dependencies]
-docopt = "0.8"
 secstr = "0.3.0"
 serde = "1.0"
 serde_derive = "1.0"
